Pull method missing logic from address to base.
Add a tiny DSL to declare flexibility on a class:

class Whatever
flexible :whatevs
end

This is encoded in the locale file as

faker:
  whatevs:
    insult: ["U suck!", "Pizzaface", "Lame"]

and can be called with Faker::Whatever.insult

For now this has only been added to Name and Address.

Fixes a bug in the :nl locale (tussenvoegsel).