Tests for user_name with integer or range argument
Adding tests of the optional argument to the Faker::Internet.user_name
method so that it assumes that if it is an integer it is specifying a
lower-bound to the length of the generated user name, and if it is a
range than it is a lower and upper bound to the length of the generated
user name.

For example:
  Faker::Internet.user_name 4
would never generate "abc" as a user name, but might generate "abcd".

Another example:
  Faker::Internet.user_name 4..8
would never generate "abc", might generate "abcd", might generate
"abcdefgh", but would never generate "abcdefghi".