Added support for email addresses in ru locale.
Previosly, Faker::Internet.email wasn't working for :ru locale, because
domain_name part of the email is based on Faker::Company.name,
which is in russian letters.

Patch is based on :uk locale implementation, which was sligtly fixed
to support locale as symbol. Fixed rare test failures in uk locale
due to cyrilic domain suffixes.