Fix HSLA colors format
Corrected the HSLA colors format. The method was returning an array of
three floats between 0 and 360 and a float between 0 and 1. The correct
format is an integer between 0 and 360, two floats to second point
precision between 0 and 1, and one float between 0 and 1 to one point
of precision. Changed variable names in HSL and HSLA methods for
clarity and updated the corresponding tests.