Correct HSL and HSLA color formatting (#768)
* Fix HSL color format

The HSL color format was returning an array of floating point numbers
between 0 and 360. The correct format is one integer between 0 and
360 followed by two percentages, in this case represented as floats
between 0 and 1. The tests have been changed to reflect this and the
method has been updated.

* Fix HSLA colors format

Corrected the HSLA colors format. The method was returning an array of
three floats between 0 and 360 and a float between 0 and 1. The correct
format is an integer between 0 and 360, two floats to second point
precision between 0 and 1, and one float between 0 and 1 to one point
of precision. Changed variable names in HSL and HSLA methods for
clarity and updated the corresponding tests.