Fixes NL postcode
* An effort was made in https://github.com/stympy/faker/commit/12df44f7e8a22ba68cdc4e90c5a771bd9d619c07
to account for Netherlands postcodes never including the letters SS, SD,
or SA. This let to some odd output, as mentioned in https://github.com/stympy/faker/issues/1188.

* The issue, as far as I can tell, is that `Faker#regexify` is built to
handle negative lookbehind regexes.

* This is a naive solution, but simple hardcodes all the allowed letter
combos in.

* As for a test case, the previously implemented one as `test_nl_postcode`
was actually an intermittent failure (which could be seen by running it
a bunch of times in a row). I've kept it as is because I'm not sure how
to make it better, short of telling it to execute a ton of times (which
seems silly). Open to ideas on that.