Use obviously invalid arguments in tests
'Artificer' and 'Goliath' are invalid options in the scope of what the
current options are, but are actual options created by Wizards of the
Coast in content not covered by the System Reference Document. These
were changed to 'Jedi' and 'Wookiee' to make it obvious that they are
invalid options.