Make locale self-contained
This is the final MR which makes the locale self-cotained. This came
out of my work at https://github.com/psibi/fakedata

I have been fixing the non working parts of the faker locales from
last year and have been documenting things which doesn't work:
https://github.com/psibi/fakedata/blob/master/Development.md#known-locale-issues

This MR fixes all of them and with this fix - the Haskell client will
become on par with the Ruby. (It still lag behinds becomes it's
tracking via git submodule - but it will completely support all the
feature for the particular submodule!)