Deprecate `HeroesOfTheStorm.class`
Fixes https://github.com/faker-ruby/faker/issues/2006.

This PR solves an issue of overriding Ruby's built-in `Object#class` method.

I'd like to maintain `HeroesOfTheStorm.class`'s behavior for
a period of time.
But I chose to make a breaking change because it differs from
the expected behavior for built-in `Object#class` method.

I think that the warning for `HeroesOfTheStorm.class`
will be removed in near future.