Remove full-stop from some titles
The UK convention is that a contraction, such as 'Mr' for 'Mister', does not
have a full-stop (period) while an abbreviation, such as 'Prof.' for 'Professor'
does. This is different from the US convention, where a period is always used.

See https://www.grammar-monster.com/lessons/abbreviations_contractions_full_stops_periods.htm