Strip extra whitespace from faker parsed output
We found that when using Faker to generate data, we'd have situations
with itermittent test failures.

The culprit was that, on occassion, Faker would generate a value that
was intened to have multiple pieces to it, but only return a string with
one of the items.

Ex:

-> coffee.yml
```yml
en:
  faker:
    coffee:
      blend_name: "#{name_1} #{name_2}"
```

```ruby
Faker::Coffee.unique.blend_name # => "Holiday Blend"
```

But sometimes we'd get a missing `name_2` value:

```ruby
Faker::Coffee.unique.blend_name # => "Holiday "
```

In this instance, we'd be checking our views to see if the `name` of the
record (using a coffee name faker) existed on the page. However, HTML
output strips whitespace from values - so the test would say:

"Find 'Holiday ' on the page", and the HTML would have the string
"Holiday" instead.

This change strips whitespace from output to ensure that, if a value is
missing from the generated fake, that there are no extra spaces.