fixes flaky tests for bot and user agents
The previous test was matching against a fixed list of options.
The problem arised when a new type of user agent was
added to the locales. Because the generator randomly picks
an user agent, the previous test was very flaky.

The generator should return a string even if
the vendor is not available or the argument is invalid.

So we can simply verify that the agent is present.
There's no need to validate it against a regex,
because we would have to keep this regex updated
whenever a new user agent is added to the locale.