Fix Vehicle VIN containing dash character
The regression was introduced in #d113aabf33 when additional RuboCop
cops were enabled which resulted in `0-0` being corrected to `0-` which
is not the same as the former denotes a range whereas the latter is
simply the chars `0` and `-`.

Looking at the Wikipedia page for plant identifiers it’s not clear what
characters position 11 should be limited to but using `0-9` is a rather
safe assumption. Though, using `[A-Z0-9]` would probably work as well.