Remove tzinfo dependency (#2952)
faker unfortunately can't be relying on gems for
validating random generated data for every generator
we have. That introduces too much work for maintainers.

Having one dependency for tests means opening the doors for more
dependencies later on. Although we do our best to keep the library
up to date, we can't keep up with all the validations to be passing.

If users rely on valid Timezones for their apps, we recommend not
relying on faker for that and instead ensure valid data is generated
on their end.