Deprecator improvements
In preparation for removing deprecations for certain generators,
this commit documents the Deprecator and how to test deprecated
generators. It also adds a unit test using newly added matchers
based on Rails deprecation matchers.

With this change, we can remove generators in a follow up PR
while having the Deprecator covered and documented.