Add deprecation warning for Faker::Twitter
Besides renaming the generator to Faker::X,
the generator is vastly outdated. Twitter, aka X API, has changed a lot (it's now v2),
and most of the attributes being returned in the user have been deprecated:
https://docs.x.com/x-api/enterprise-gnip-2.0/fundamentals/data-dictionary#no-longer-supported-deprecated-attributes

To maintain backwards compatibility, users still using Faker::Twitter
will keep using it but be notified of the upcoming changes.

Initially, the goal was to only update the `user` generator but
then I realized it wouldn't make sense to leave `status` as it is.