Deprecate `Faker::Twitter` in favor of `Faker::X` and update attributes (#3138)
* Add deprecation warning for Faker::Twitter

Besides renaming the generator to Faker::X,
the generator is vastly outdated. Twitter, aka X API, has changed a lot (it's now v2),
and most of the attributes being returned in the user have been deprecated:
https://docs.x.com/x-api/enterprise-gnip-2.0/fundamentals/data-dictionary#no-longer-supported-deprecated-attributes

To maintain backwards compatibility, users still using Faker::Twitter
will keep using it but be notified of the upcoming changes.

Initially, the goal was to only update the `user` generator but
then I realized it wouldn't make sense to leave `status` as it is.

* Deprecate Faker::Twitter and update status to tweet

After working on the user generator, it made sense to also deprecate
the `status` one. The API migration docs don't mention changing status to tweet,
but after inspecting what `status` returns, I realized the new API calls is `tweet`.

There were lots of changes for both user and tweet objects. Most notably, the shape of the response,
additional fields, and removal of deprecated attributes. Used resources as references:

- https://docs.x.com/x-api/migrate/data-format-migration#user-object
- https://docs.x.com/x-api/fundamentals/data-dictionary#tweet
- https://docs.x.com/x-api/migrate/data-format-migration#entities-and-expanded-entities-objects

* Update docs